Restructured billboard-state to follow file-structure in project
