+++ app/(dashboard)/(routes)/[storeID]/billboards/page.tsx
@@ -5,7 +5,7 @@ import { columns } from "@/components/stores/dashboard/billboards/columns";
 import { useEffect } from "react";
 import { PostForm } from "@/components/stores/dashboard/billboards/PostForm";
 import axios from "axios";
-import { BillboardState } from "@/components/stores/dashboard/billboards/state";
+import { BillboardState } from "@/components/stores/dashboard/billboards/zustand/zustandstate";
 import { useParams } from "next/navigation";
 import { APIList } from "@/components/stores/dashboard/api-list/APIList";
