+++ components/stores/dashboard/billboards/deleteBillboard.tsx
@@ -1,17 +1,17 @@
 import axios, { isAxiosError } from "axios"
 import toast from "react-hot-toast";
-import { BillboardState } from "@/components/stores/dashboard/billboards/state";
+import { BillboardState } from "@/components/stores/dashboard/billboards/zustand/zustandstate";
 import { useParams } from "next/navigation";
 
 export function MyDelete(params: { itemId: string }) {
   const reFetchBillboards = BillboardState((state) => state.reFetchBillboards);
-  const {storeID}= useParams()
+  const { storeID } = useParams()
   return (
     <div onClick={async (e) => {
       axios.delete(`${process.env.NEXT_PUBLIC_BACKEND_URL}/api/${storeID}/billboards/${params.itemId}`)
         .then(function (response) {
           toast.success("Billboard deleted successfully");
-          reFetchBillboards(Array.isArray(storeID)?storeID.toString():storeID)
+          reFetchBillboards(Array.isArray(storeID) ? storeID.toString() : storeID)
           return response
         })
         .catch(function (error) {
